refactor: improve package structure
